+++ lib/helpers/app_helpers.js
@@ -172,8 +172,8 @@ function app_helpers() {
 
     //generic message for logging and console writing
     write_message: function(level, msg, val){
-      console.error(msg + ': ' + val);
-      logger.log(LOG_LEVEL_ERR, msg + ': ' + val);
+      console.error(msg + ':', val);
+      logger.log(LOG_LEVEL_ERR, msg + ':', val);
     },
 
   }
