+++ lib/usgs_api/usgs_helpers.js
@@ -130,7 +130,7 @@ module.exports = {
           // of error in the returned from USGS JSON API
           if (response_data === null || response.data.error){
             var error =  self.get_response_error(response);
-            this.throw_error(error)
+            return this.throw_error(error)
           }
 
           return response_data;
